# ⚡ 🐛 Fix `Integer` decoding bug and optimize