In terms of performance, improvements are pretty minor. 🙄
However, there were bugs regarding decoding where invalid formats were allowed.  See the newly added test.
Cases 1, 3, and 4 fail on 0.13.0.

# Benchmark results

Tested with 100,000 encoded `Integer`s. `DecodeSmall` was generated using `Random.Next()` and `DecodeBig` was generated using `Random.NextInt64()`.

## 0.13.0

|      Method |     Mean |    Error |   StdDev |       Gen0 | Allocated |
|------------ |---------:|---------:|---------:|-----------:|----------:|
| DecodeSmall | 30.88 ms | 0.603 ms | 0.922 ms |  8500.0000 |  50.86 MB |
|   DecodeBig | 50.19 ms | 0.485 ms | 0.405 ms | 11500.0000 |  70.18 MB |

## Updated

|      Method |     Mean |    Error |   StdDev |      Gen0 | Allocated |
|------------ |---------:|---------:|---------:|----------:|----------:|
| DecodeSmall | 26.09 ms | 1.067 ms | 1.897 ms | 5500.0000 |  33.98 MB |
|   DecodeBig | 46.98 ms | 0.912 ms | 1.420 ms | 8750.0000 |   53.4 MB |